chore: CI workflow to only upload report if CODACY_PROJECT_TOKEN is set
GitHub will not set secrets in the environment when a pull request is submitted
from a forked repository. This commit modifies the CI workflow to only send
the report when the API token has been set. It also consolidates the two
workflows into a single one which shares coverage data between jobs.

Fixes: https://github.com/cloudevents/sdk-javascript/issues/190

Signed-off-by: Lance Ball <lball@redhat.com>