chore(actions): don't auto-close stale issues and pull requests
Open issues and pull requests should be closed by a human who has considered
the issue/pr and made a choice based on the content of the issue/pr and the
state of the repository.

What this means in practice is that some issues will stay open for a long
time, even if they are stale. They may be open because it really is something
that will be considered for future work. Periodic issue triage should be
done by the maintainers, and if an issue _should_ be closed - for example,
the issue is no longer relevant - then a maintainer can close it.

Signed-off-by: Lance Ball <lball@redhat.com>