fix: improve binary data detection in HTTP transport
Previously, we were only checking for a single typed array,
`UInt32Array` but that's not quite broad enough, since binary
data can exist in various forms. This means of checking for
binary data, using `types.isTypedArray` is not foolproof, but
is better than previous, as the previous implementation would
not recognize a PNG file read from disk via `fs.readFileSync`
as binary data. Now it will.

Signed-off-by: Lance Ball <lball@redhat.com>